requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# --- CORE DATA STACK ---
polars>=0.20.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0
pyarrow>=14.0.0      # Critical for Parquet file I/O
tabulate>=0.9.0      # For formatted table output

# --- LIVE TRADING & HARVESTING ---
fyers-apiv3>=3.0.0
websocket-client>=1.6.0
selenium>=4.15.0     # For headless browser login
webdriver-manager>=4.0.0
pyotp>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0     # For Fyers token refresh API
boto3>=1.34.0        # AWS SDK for SSM/Secrets Manager (cloud deployment)

# --- SIMULATION & ML (ABIDES Dependencies) ---
# Note: ABIDES itself is installed from source (see INSTALL_GUIDE.md)
matplotlib>=3.7.0
scipy>=1.10.0

# --- DEEP LEARNING (CPU) ---
torch>=2.1.0
torchvision>=0.16.0